fix bigint for chatId

+++ Database/001-Tables/001-Create/dbo.Chats.sql
@@ -1,7 +1,7 @@
 IF OBJECT_ID('[dbo].[Chats]') IS NULL
     create table dbo.Chats
     (
-        id int
+        id BIGINT
             constraint Chats_pk
                 primary key nonclustered,
         language VARCHAR(5)  not null,
